// frontend/src/api-generated/models/phongBanResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface PhongBanResponse {
  id?: number;
  idDonVi?: number;
  maPhongBan?: string;
  tenPhongBan?: string;
  tenTiengAnh?: string;
  tenVietTat?: string;
  soMayLe?: string;
  soHotlinePhong?: string;
  emailNhom?: string;
  loaiPhongBan?: string;
  hanMucNganSach?: number;
  maTrungTamChiPhi?: string;
  moTaChucNang?: string;
  trangThai?: string;
  thoiGianThanhLap?: string;
}